Добавление метода PUT

+++ HomeApi.Data/Repos/DeviceRepository.cs
@@ -97,7 +97,11 @@ namespace HomeApi.Data.Repos
         /// </summary>
         public async Task DeleteDevice(Device device)
         {
-            _context.Devices.Remove(device);
+            var entry = _context.Entry(device);
+            if (entry.State == EntityState.Detached)
+                _context.Devices.Remove(device);
+
+            // Сохранение изменений
             await _context.SaveChangesAsync();
         }
     }
